KUDU-1769: fs check action with rudimentary data block GC
This commit introduces a new "fs check" action to the CLI tool. For the time
being its capabilities are slim:
- Detecting missing blocks.
- Detecting orphaned blocks, optionally "repairing" (i.e. deleting) them.

I'm sure it'll be augmented to perform more consistency checks and repairs
in the future.

Why build this in the CLI tool and not in the block manager? A few reasons:
1. This "poor man" approach to GCing orphaned data blocks is extremely
   heavy-weight. Doing it at server startup is probably a bad idea.
2. Performance aside, finding a good time at server startup to compare the
   block lists will require additional refactoring and reordering.
3. If we change our minds, we can always move it later.

I do expect that the block manager will take on some check or repair duties,
but they'll be relegated to startup time, and will be much smaller in scope
(e.g. the LBM could repunch holes that previously failed to punch).

Change-Id: I82da8e234c338cd5a7540a22cebfa0a4958388ec
Reviewed-on: http://gerrit.cloudera.org:8080/6361
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>
Reviewed-by: Dinesh Bhat <dinesh@cloudera.com>