tpch: allow hash partitioning
This enables a hash-partitioning mode for tpch_real_world. In this mode,
we still create one tablet per thread, but we use hash-partitioning so
that we don't get the perfect one-to-one insert pattern that we get with
range partitioning.

This is a bit more realistic for many batch insert scenarios, in which
many writers write locally-sorted (but overlapping) data ranges into a
single hash-partitioned table.

Change-Id: Icbbb643447dff58b32e67764f6fb632441a4f6e4
Reviewed-on: http://gerrit.cloudera.org:8080/6709
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>