[consensus] Add consensus op-level lag metrics
It may be useful to see how far behind the leaders each follower is
with regard to replicating operations. As such, this patch adds a
metric facilitate this.

The new metric counts the number of operations behind the leader each
peer is. It is updated when a follower receives a request from a
leader, and after the peer message queue appends operations.

Change-Id: Ida8e992cc2397ca8d5873e62961a65f618d52c36
Reviewed-on: http://gerrit.cloudera.org:8080/6451
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins