generic_iterators: avoid holding completed iters in merge/union
Previously, the MergeIterator and UnionIterator would hold on to all of
the iterators they consume until the entire process completed. Each such
iterator may hold on to blocks in the block cache (eg dictionary blocks)
and consume a fair amount of memory.

Without this patch, I noticed that issuing a scan caused the block cache
consumed memory to slowly increase until the scan finished, at which
point it would quickly drop. With the patch, I no longer see this
behavior (memory usage is constant).

The one feature lost here is the stringification of MergeIterator which
previously included all of the sub-iterators. We only used this in one
place in a VLOG(1) message anyway, and for typical merges across
thousands of DRS iterators, this wasn't likely to be particularly useful
anyway.

Along the way, I did a bit of cleanup in cfile_set.cc to use unique_ptr
instead of pre-C++11 raw pointers.

Change-Id: I8e068c256c5855efeff5c6325dbb5fbf770eb900
Reviewed-on: http://gerrit.cloudera.org:8080/6460
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <dralves@apache.org>