interval_tree: allow bulk queries
This adds an API to interval_tree to allow it to be queried in "bulk".
The caller passes a set of query points and a callback, and the tree
calls the callback once per <interval, point> pair.

This has much better CPU cache efficiency than a lot of separate
queries. Additionally, doing 'Q' queries in bulk is logarithmic in Q
rather than linear.

More importantly, this will allow rowset-wise processing of batches of
writes, which opens up big perf wins in later patches in this series.

Change-Id: Ifb4da25ca43413fbcae631a7b0f3f16062e4e408
Reviewed-on: http://gerrit.cloudera.org:8080/6481
Tested-by: Todd Lipcon <todd@apache.org>
Reviewed-by: David Ribeiro Alves <dralves@apache.org>