rowset_metadata: use flat_map for column->block mapping
In a server with around 1T of data, the std::map<> here was taking
around 25MB of RAM.

Based on an experiment, a map<int,int> with 40 entries takes 1648 bytes
whereas a flat_map<int,int> (after shrink_to_fit) takes 344 bytes. So,
this should save around 20M of RAM from that server (and improve memory
locality while we're at it).

Change-Id: I1b7eaac3481672c69709009b6eacc2008027d4cd
Reviewed-on: http://gerrit.cloudera.org:8080/6470
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Todd Lipcon <todd@apache.org>