write_transaction: allocate RowOp objects from the arena
This improves the memory locality of write transactions by making all of
the RowOps in a batch come from the transaction's arena. Additionally
this reduces contention on tcmalloc since the arena grabs memory in
bulk.

I didn't benchmark this before/after, but in various write stress
testing this made a bit of a difference in cache miss rates and CPU time
spent in RowOp code.

Change-Id: I8e7969423c057f9bf1169fe066819228c59ed269
Reviewed-on: http://gerrit.cloudera.org:8080/6480
Reviewed-by: David Ribeiro Alves <dralves@apache.org>
Tested-by: Kudu Jenkins