docs: Add a section for evolving gflags
This patch adds a new section for evolving gflags in the documentation
and changes some of the XSLT logic to be more consistent.

It would be nice to replace the XSLT with a Python script.

Change-Id: I791d91e4f7ace4483958d26a60817857899fb474
Reviewed-on: http://gerrit.cloudera.org:8080/6489
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
(cherry picked from commit 824b0f46dc3d1443ad10db1b07cb085cb5bd7a4e)
Reviewed-on: http://gerrit.cloudera.org:8080/6546
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>