docs: Move documentation style guide from user guide to design-docs
This isn't very important to the typical end-user and doesn't really
belong in the user guide. Let's move the way it's primarily viewed from
the web site to the git repo and link people to the GitHub rendering of
the doc from the contributing page of the user guide.

Change-Id: Ia9caa8767e2b3dd26cac11157b80a0b452d830c4
Reviewed-on: http://gerrit.cloudera.org:8080/6490
Tested-by: Mike Percy <mpercy@apache.org>
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>
(cherry picked from commit 58e1d26d831a25835eb83aa21ac964f48792fdfb)
Reviewed-on: http://gerrit.cloudera.org:8080/6547
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>
Tested-by: Jean-Daniel Cryans <jdcryans@apache.org>