KUDU-1962. Fix NPE when master returns SERVER_TOO_BUSY
If the master's RPC queue overflows, it returns a TOO_BUSY error to the
clients. We were not properly handling this in the case of
ConnectToMaster RPCs.

This adds a new unit test which provokes the behavior by setting a short
queue and injecting latency into tablet location lookups while starting
several threads which make clients and perform lookups. This also fixes
the issue by reverting back to passing the placeholder 'masterTable'
into these RPCs.

Longer term we should avoid this 'masterTable' hack, since the master is
not a table, but given that this causes serious problems, a more
targeted fix is desirable for 1.3.1 and 1.4.0.

Change-Id: Ic4aba2e39f05488da500c18247c72181c8e9dceb
Reviewed-on: http://gerrit.cloudera.org:8080/6516
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>
Tested-by: Kudu Jenkins
(cherry picked from commit bee8e875e21acb1d36f104f8388fc9ed6c963821)
Reviewed-on: http://gerrit.cloudera.org:8080/6519