KUDU-1963. Avoid NPE and SSLException when a TabletClient is closed during negotiation
This fixes two related bugs which were triggered when a client called
close() on a TabletClient (e.g. because the end user of the API called
KuduClient#close()):

1) javax.net.ssl.SSLException: renegotiation attempted by peer; closing the connection

This was caused by the following interleaving in Netty:

T1                                          T2
------------------------                    -----------------
TabletClient.close()
  Channel.close()
    SslHandler.close()
      - starts a TLS "shutdown" message
      - underlying SSLEngine is now in a
        "closing" state                     call Channel.write()
                                               call SSLEngine.write()
                                                 - gets NEEDS_UNWRAP
                                                   result
                                                 - interprets this as a
                                                   renegotiation and
                                                   throws an Exception

      - call SSLEngine.unwrap() to get the
        TLS message data to send

I wasn't able to write a unit test that triggered this reliably, since
TabletClient is too tightly coupled to ConnectionCache and
AsyncKuduClient. However, I was able to reproduce this reliably on an
Impala cluster by starting 100 threads sending short queries round-robin
to 9 impalads. Previously I saw the exception within a few seconds and
now I've run successfully for many minutes with no errors.

2) NullPointerException following above SSLException

When the above SSLException occurred, it called the
TabletClient#exceptionCaught() method, which called 'cleanup()' to fail
any outstanding RPCs. However, because the exception was being caught
because the client had already been explicitly closed, the
'rpcsInflight' and 'pendingRpcs' lists had already been set to null.
This caused an NPE as we tried to process those lists for a second time.

Another case which cased a similar NPE was that the NegotiationResult
could come after the close() had been called, in which case
'pendingRpcs' would already be null.

This patch fixes both issues as follows:
- if we've already explicitly closed the client, we ignore SSLExceptions
  rather than logging them. Actually avoiding this race seemed to be
  extremely complicated, so better to just ignore it.
- properly check whether the pendingRpcs/rpcsInFlight lists were already
  null before processing them.

Change-Id: Ide91722446d01acb58e3e181e0d10932e023a043
Reviewed-on: http://gerrit.cloudera.org:8080/6541
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>