[security-faults-itest] fix test flakiness
When running the security-faults-itest on a slow VM with other
activities in parallel, it might take up to 8 seconds to start a master.
So, bumping the ticket TTL to higher value so Kudu cluster with 3
masters finishes starting up when initially acquired tickets are valid.

Also, updated Raft parameters interval to speed up leader election.

Change-Id: I132426ee37358a228cf3f5d1fef9e9ec08610b16
Reviewed-on: http://gerrit.cloudera.org:8080/6586
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>