KUDU-1965: Allow user provided TLS certificates to work with KRPC
This patch adds a bool in the TlsContext class to keep track of
whether the certificates provided are externally provided or not.

If it is externally provided, then authenticated TLS is not
negotiated, causing a fallback to SASL authentication. The
certificates are used only for encryption.

This adds some tests in negotiation-test

Change-Id: Idd44770a1fe85e9934a657ee79d93139b9a86dff
Reviewed-on: http://gerrit.cloudera.org:8080/6594
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Kudu Jenkins