[mini-kdc] /dev/stderr instead of STDERR for kdc logs
This is a workaround for kdc5krb trying to open STDERR in read-write
mode "a+" (krb5/src/lib/kadm5/logger.c) in versions prior to 1.13.

Since mini-kdc is run via the SubProcess utility, its STDERR is piped
via write end of the pipe and cannot be opened in read mode.  That leads
to error messages like:

  krb5kdc: cannot parse <STDERR>
  krb5kdc: warning - logging entry syntax error

Specifiying 'FILE:/dev/stderr' allows for using the same destination
for the error output, but not opening the standard error in read mode.

NOTE: the issue in logger.c has been addressed by changelist a8592307
      in Kerberos repository and the fix is included in krb5 starting
      version 1.13.

Change-Id: I212e2cceb80acbe88040b16d9055cf01ac8a761b
Reviewed-on: http://gerrit.cloudera.org:8080/6602
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>