[docs] Add admin workflow for recovering from disk failure
I didn't document how to rebalance tablets onto the repaired tserver if
necessary, since the process is complicated and error prone, and we hope
to have a rebalancing tool in the future. These docs will quickly become
outdated when KUDU-616 is fixed, but I think it's worth it to document
since we frequently receive questions on the topic.

Change-Id: I6541bffc5e9546c523df610fd8c025dd05e403bf
Reviewed-on: http://gerrit.cloudera.org:8080/6606
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Andrew Wong <awong@cloudera.com>
(cherry picked from commit 87154f4a39c77ab92d80f3effa58de3000921127)
Reviewed-on: http://gerrit.cloudera.org:8080/6677
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>
Tested-by: Dan Burkert <danburkert@apache.org>