log: shut down appender thread when idle
This changes the log appender thread to be based on submitting tasks to a pool,
rather than a thread which is always running.

See the comments on Log::AppendThread in log.cc for notes on the design.

This also adds a new test case in mt-log-test which triggers the various
complexities of the appender lifecycle. I manually checked coverage using
LOG(WARNING) statements to make sure it does exercise the various cases.

Change-Id: Id40a4cfcec96198b537c2f50be7ff1204faf96d2
Reviewed-on: http://gerrit.cloudera.org:8080/6856
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@apache.org>