Tweak ListMasters RPC error handling
Unlike other master-specific RPCs, the ListMasters response contains an
AppStatusPB error type instead of a MasterErrorPB.  This makes it harder
to use ListMasters in a generic context with other master RPCs.

This commit changes the error type, while preserving the old type as a
deprecated field. There are no known users of the ListMasters RPC yet,
but maintaing backwards compat is still part of our contract.

Change-Id: I28e175df2cecad4f62806bc3ea0c314e5b861e40
Reviewed-on: http://gerrit.cloudera.org:8080/6704
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>