gutil: stop supporting AMD Opteron K8 (take 2)
This removes a workaround for a bug in very old (circa 2003) Opteron
chips. The workaround was producing an extra branch around most of our
atomic operations. This likely has a negligible perf impact, but it's
possible the slightly smaller code would make us run a bit faster or
reduce branch mispredictions.

Previously we also entirely forgot to call this function, so it was in
fact never applying the workaround. This patch marks the function as a
constructor, which ensures that it gets automatically called by any
binaries using atomic ops.

Impala made a similar change[1][2] a while back and seems to have had
no real issues.

[1] https://gerrit.cloudera.org/#/c/2516/
[2] https://github.com/apache/incubator-impala/commit/a7c3f301bb1da14d6bb35d37aab47450c67cca0c

Change-Id: I4eff19c62741e7433e378f7e0725079ecbcf1a97
Reviewed-on: http://gerrit.cloudera.org:8080/6680
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins