gutil: stop supporting AMD Opteron K8
This removes a workaround for a bug in very old (circa 2003) Opteron
chips. The workaround was producing an extra branch around most of our
atomic operations. This likely has a negligible perf impact, but it's
possible the slightly smaller code would make us run a bit faster or
reduce branch mispredictions.

This also adds a missing call to init the feature flags in the first
place, which we seem to have forgotten the first time around.

Impala made a similar change[1][2] a while back and seems to have had
no real issues.

[1] https://gerrit.cloudera.org/#/c/2516/
[2] https://github.com/apache/incubator-impala/commit/a7c3f301bb1da14d6bb35d37aab47450c67cca0c

Change-Id: I59b5aeaae56b5f310b5dd160b868fcb5d04f4da1
Reviewed-on: http://gerrit.cloudera.org:8080/6661
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins