[cmake] fixed 'latest' dir symlink on MacOS X
If the 'latest' symbolic link already exists, do not follow it, so the
'latest' link is updated to point to the new src directory.  Prior to
this, the result symlink on OS X would be created in the directory
which the 'latest' points at.

The fix is in replacing Linux-specific '-T' option with '-n' option
which is compatible at least between BSD-style and Linux ln
implementations and adding that '-n' flag regardless of the platform.

Change-Id: Ica6314eb5274e853b7b39b7e6ede17305d0c646d
Reviewed-on: http://gerrit.cloudera.org:8080/6667
Reviewed-by: Dan Burkert <danburkert@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>