KUDU-1826 add propagated timestamp to sync client
Updating and getting the propagated timestamp only exist as functions
of the AsyncKuduClient, but these functions are still useful to the
KuduClient.

This patch exposes AsyncKuduClient's updateLastPropagatedTimestamp()
and getLastPropagatedTimestamp() to the KuduClient. The static value
NO_TIMESTAMP is also added to the KuduClient so timestamps returned
by KuduClient.getLastPropagatedTimestamp() can be compared with
KuduClient.NO_TIMESTAMP.

Change-Id: Ibbfeb9526423f5acd0ef6482869d36cdd146e921
Reviewed-on: http://gerrit.cloudera.org:8080/6798
Reviewed-by: David Ribeiro Alves <davidralves@gmail.com>
Tested-by: David Ribeiro Alves <davidralves@gmail.com>